one cleanup pass
